+++ src/index.ts
@@ -1,16 +1,4 @@
-import { EventEmitter, Subscription } from "expo-modules-core";
-
-import ReactNativeTurboUtilsModule from "./ReactNativeTurboUtilsModule";
 import { Contact, DeviceInfoType, InstalledApps } from "./types";
+import * as ReactNativeTurboUtilsModule from "./load-module";
 
-export { Contact, InstalledApps, DeviceInfoType };
-
-const emitter = new EventEmitter(ReactNativeTurboUtilsModule);
-
-export function addDeviceInfoChangedListener(
-  listener: (event: DeviceInfoType) => void
-): Subscription {
-  return emitter.addListener("onDeviceInfoChanged", listener);
-}
-
-export default ReactNativeTurboUtilsModule;
+export { Contact, InstalledApps, DeviceInfoType, ReactNativeTurboUtilsModule };
