--- src/index.web.ts
@@ -1,4 +1,0 @@
-// This is a no-op file, as we do not have any role on the web
-import { Contact, DeviceInfoType, InstalledApps } from "./types";
-
-export { Contact, InstalledApps, DeviceInfoType };
